Adding param assertion behavior
Rather than make the param value equality assertion from a method on the
application, I'm implementing the logic in a method on the param itself.
This allows us to store (a pointer) to the reference on the param which
can later be used for reporting.

The assert method introduced in this commit mutates the parameter
object, so it's suffixed with an exclamation point. Risk and confusion
is reduced by tracking if the param has been asserted upon previously,
and raising an exception if an attempt is made to assert again.