Removing mutant resource type behavior
I'm concerned that the aesthetic draw of defining the attributes to be
asserted on the assertion resource itself is not worth the risk involved
with such a deep coupling with the Puppet internals.

This commit removes the coupling, and adds an attributes parameter which
is intended to contain a hash of attribute/value pairs which will be
asserted against the subject.