Changing application assertion behavior
This commit sets the stage to move assertion behavior to the type.
Instead of calling its own assertion method, the application now calls
a class method on the type, passing the resource instance. While likely
this is unneeded abstraction, I feel better about the behavior being a
concern of the type rather than the application.

This commit also adds catalog finalization which is required to trigger
the munge and validate methods on the type params.