Only allow a single attribute and expectation
The assertion resource type now allows for a subject, attribute, and
expectation. During evaluation, the value of the subject's attribute
will be asserted for equality against the given expectation.

I feel that this is more idiomatic to Puppet, as a single logical
assertion now maps to a single assertion resource.