Allow arbitrary class params
This commit implements the stub_type's stub resource type on the
stub_class method, in order to allow the stubbed class to have any
param keys assigned values.

This behavior is covered in the acceptance test, but we should probably
look at covering in integration also, though I'm not 100% sure what that
would look like in the current test design.