Refactoring the application to use the reporter
This commit almost entirely redesigns the spec application to the end
that the assertion resources are evaluated by the reporter, which is a
stateful object used over the span of application execution.

In the process, the assertion result output was altered slightly, mostly
around the expected and received values which are now rendered by the
Hash.to_s method.