Removing subject type validation
I actually don't care if a user gives a non-resource reference value for
subject. If it's a string, the resource of a matching slug will be
pulled from the catalog and evaluated as expected. If it's not in the
catalog, a mystical error will be raised which will be addressed in the
future.