Rethink logic behind determination of a resource's state in the catalog
This commit removes the assumption that the application will set the
subject, and evaluates the subject itself for its presence in the
catalog.

This is largely due to the fact that we can't set the value of subject
to a symbol from the application due to the (necessary) type validation,
and because we still need the resource slug for use in the error output.